move metadata apps into nie-OS-apps folder
