+++ src/app/development-views/metadata-view/metadata-view.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MetadataViewComponent } from './metadata-view/metadata-view.component';
 import { RouterModule } from '@angular/router';
-import { FactSheetModule } from '../../fact-sheet/fact-sheet.module';
+import { FactSheetModule } from '../../nie-OS-apps/fact-sheet/fact-sheet.module';
 import { MaterialModule } from '../../material.module';
-import { TagChipsModule } from '../../tag-chips/tag-chips.module';
+import { TagChipsModule } from '../../nie-OS-apps/tag-chips/tag-chips.module';
 import { ArithmeticModule } from 'nie-ine';
 import { D3jsModule } from '../../nie-OS/apps/d3js/d3js.module';
