+++ src/app/development-views/metadata-view/metadata-view/metadata-view.component.spec.ts
@@ -1,9 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { MetadataViewComponent } from './metadata-view.component';
-import { FactSheetModule } from '../../../fact-sheet/fact-sheet.module';
+import { FactSheetModule } from '../../../nie-OS-apps/fact-sheet/fact-sheet.module';
 import { MaterialModule } from '../../../material.module';
-import { TagChipsModule } from '../../../tag-chips/tag-chips.module';
+import { TagChipsModule } from '../../../nie-OS-apps/tag-chips/tag-chips.module';
 import { RouterTestingModule } from '@angular/router/testing';
 import { By } from '@angular/platform-browser';
 import { KnoraV1RequestService } from '../../../shared/knora/knora-v1-request.service';
