+++ src/app/nie-OS-apps/fact-sheet/fact-sheet.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FactSheetComponent } from './fact-sheet/fact-sheet.component';
 import { RouterModule } from '@angular/router';
-import { BaseTypeFormsModule } from '../nie-OS-apps/shared/base-type-forms/base-type-forms.module';
+import { BaseTypeFormsModule } from '../shared/base-type-forms/base-type-forms.module';
 
 @NgModule({
   imports: [
